### 2016-11-13 #94 @labarba
We're looking for volunteers to review this submission to JOSS. It is a **sci-kit-learn** compatible project. Any chance that we could get a review from @amueller, @GaelVaroquaux or maybe @NelleV?
