### 2016-11-13 #94 @arfon
> The message I take from this is that I should make sure the papers I'm assigned are reviewed slowly so I don't get assigned more. Not sure this is quite right...

@danielskatz - `¯\_(ツ)_/¯` I guess so? Currently there's little/no visibility into who is reviewing what....and we've got 8 submissions currently without an editor. Just trying out some tooling to help make sure we have a relatively even load on each editor.

Any suggestions for improvements would be very welcome :-)

> Topic-wise, not in my area of expertise. I could activate some contacts to see if a reviewer can be located, but would not be able to moderate if author and reviewer disagree.

Understood. Would you be willing to start editing this and if there's a disagreement then perhaps we can bring in a domain expert?
