### 2016-11-13 #94 @danielskatz
Maybe we should also look at how many papers we have edited in the last 3 and 12 months (yes, I know 12 months won't say different, *yet*)
