### 2016-11-13 #94 @arfon
> Maybe we should also look at how many papers we have edited in the last 3 and 12 months (yes, I know 12 months won't say different, yet)

Yeah. That's a good point. I'll look at adding that.
