### 2016-11-13 #94 @GaelVaroquaux
> It is a sci-kit-learn compatible project. Any chance that we could get
> a review from @amueller, @GaelVaroquaux or maybe @NelleV?

I really don't have the time to do everything that I should be
doing. And, to be honest, if I find time, I'd rather review PRs in
scikit-learn, where we have a terrible backlog.

Sorry :$
