### 2016-11-13 #94 @danielskatz
Thinking about the general editor balance issue a bit more...

I don't think our goal should be to balance work across editors.  We should be assigning reviews to editors based on topics.  If an editor is getting too many submission in a given topic, we should add another editor in that topic.

I think the PeerJ CS model is what we should be following here, add editors as needed, and when there are papers where the assignment to an editor is not clear, ask the editors for volunteers.
