### 2016-11-24 #94 @pjotrp
@DannyArends would you care to review?
