### 2016-11-25 #124 @whedon
[REVIEW]: Confidence Intervals for Random Forests in Python
**Submitting author:** @kpolimis (<a href="http://orcid.org/0000-0002-3498-0479">Kivan Polimis</a>)
**Repository:** <a href="https://github.com/scikit-learn-contrib/forest-confidence-interval" target ="_blank">https://github.com/scikit-learn-contrib/forest-confidence-interval</a>
**Version:** .1

- [x] **Version:** Does the release version given match the GitHub release (.1)?
