### 2016-11-25 #94 @arfon
Closing as the review is now happening over in #124
