### 2016-11-26 #124 @DannyArends
Review of scikit-learn-contrib/forest-confidence-interval

I checked the initial boxes, the boxes I left unchecked have issues and are described point by point below:

**License:**
LICENCE file is not an OSI approved licence

**Installation:**
I couldn't install one of the dependencies, without root rights,
please describe that this might be needed in the documentation.

**Functionality:**
The software seems to provide the functionality as described

**Performance:**
The software makes no performance claims, a performance comparison against
the original R-code would be interesting to add

**A statement of need:**
There is no clear statement of need is not provided in the readme

**Example usage:**
1) The examples use two standard datasets from another package, however it is not explained in which
format the users own data needs to be presented before it can be used.
Please provide links to the relevant documentations.

2) The two examples do not show before and after plots of the dataset, as
such it is not clear why the user should use this routine.

3) There is no description of what settings can be used in the calculation of inbag and unbiased variance and how these influence the results

**Functionality documentation:**
This package provides 2 functions to the user, which have been documented.
However there are also 2 undocumented internal functions, which should be documented.

**Automated tests:**
There is no mention of tests, or how to run the tests in the documentation.
Also the level of test coverage for this package is unknown

**Community guidelines:**
PLease provide guidelines in the readme file for others to
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support

**Authors:**
The submitting author: Kivan Polimis, has no institute listed, while the others from the same university have an institute listed. Is this an oversight or intentional ?

**A statement of need:**
There is no clear statement of need is not provided in the paper,

**References:**
Both references (wager_confidence_2014 and wager_randomforestci_2016) are missing a DOI
