### 2016-11-26 scikit-learn-contrib/forest-confidence-interval#29 @DannyArends
JOSS Review "Performance" [Minor]
The software makes no performance claims, a performance comparison against the original R-code would be interesting to add
