### 2016-11-26 scikit-learn-contrib/forest-confidence-interval#30 @DannyArends
JOSS Review "A statement of need" [Major]
There is no clear statement of need is not provided in the README.md or in the paper
