### 2016-11-26 scikit-learn-contrib/forest-confidence-interval#31 @DannyArends
JOSS Review "Example usage" [Major]
1) The examples use two standard datasets from another package, however it is not explained in which
format the users own data needs to be presented before it can be used.
Please provide links to the relevant documentations.

2) The two examples do not show before and after plots of the dataset, as
such it is not clear why the user should use this routine.

3) There is no description of what settings can be used in the calculation of inbag and unbiased variance and how these influence the results
