### 2016-11-26 scikit-learn-contrib/forest-confidence-interval#32 @DannyArends
JOSS Review "Functionality documentation" [Minor]
This package provides 2 functions to the user, which have been documented. However there are also 2 undocumented internal functions, which should be documented.
