### 2016-11-26 scikit-learn-contrib/forest-confidence-interval#34 @DannyArends
JOSS Review "Community guidelines" [Blocker]
Please provide guidelines in the README.md file for others to
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support
