### 2016-11-26 scikit-learn-contrib/forest-confidence-interval#36 @DannyArends
JOSS Review "References" [Blocker]
Both references (wager_confidence_2014 and wager_randomforestci_2016) are missing a DOI
