### 2016-12-15 5ff6427
RF: Induce the inbag from the forest.
