### 2016-12-15 6e8144f
RF: Inbag calculation optional, and done only with bootstrapped forests.
Test that a non-bootstrapped forest would raise an error.
