+++ forestci/forestci.py
@@ -41,6 +41,11 @@ def calc_inbag(n_samples, forest):
     Columns are individual trees. Rows are the number of times a sample was
     used in a tree.
     """
+    if not forest.bootstrap:
+        e_s = "Cannot calculate the inbag from a forest that has "
+        e_s = " bootstrap=False"
+        raise ValueError(e_s)
+
     n_trees = forest.n_estimators
     inbag = np.zeros((n_samples, n_trees))
     sample_idx = []
@@ -73,7 +78,7 @@ def _bias_correction(V_IJ, inbag, pred_centered, n_trees):
     return V_IJ_unbiased
 
 
-def random_forest_error(forest, X_train, X_test):
+def random_forest_error(forest, X_train, X_test, inbag=None):
     """
     Calculates error bars from scikit-learn RandomForest estimators.
 
@@ -84,9 +89,16 @@ def random_forest_error(forest, X_train, X_test):
     ----------
     forest : RandomForest
         Regressor or Classifier object.
-        
+
     X : ndarray
         An array with shape (n_sample, n_features).
+
+    inbag : ndarray (optional)
+        The inbag matrix that fit the data. If set to `None` (default) it
+        will be inferred from the forest. However, this only works for trees
+        for which bootstrapping was set to `True`. That is, if sampling was
+        done with replacement. Otherwise, users need to provide their own
+        inbag matrix.
 
     Returns
     -------
@@ -107,7 +119,8 @@ def random_forest_error(forest, X_train, X_test):
        Random Forests: The Jackknife and the Infinitesimal Jackknife", Journal
        of Machine Learning Research vol. 15, pp. 1625-1651, 2014.
     """
-    inbag = calc_inbag(X_train.shape[0], forest)
+    if inbag is None:
+        inbag = calc_inbag(X_train.shape[0], forest)
     pred = np.array([tree.predict(X_test) for tree in forest]).T
     pred_mean = np.mean(pred, 0)
     pred_centered = pred - pred_mean
