+++ forestci/tests/test_forestci.py
@@ -28,6 +28,16 @@ def test_random_forest_error():
     V_IJ_unbiased = fci.random_forest_error(forest, X_train, X_test)
     npt.assert_equal(V_IJ_unbiased.shape[0], y_test.shape[0])
 
+    # We cannot calculate inbag from a non-bootstrapped forest. This is because
+    # Scikit-learn trees do not store their own sample weights. If you did This
+    # some other way, you can still use your own inbag
+    non_bootstrap_forest = RandomForestRegressor(n_estimators=n_trees,
+                                                 bootstrap=False)
+
+    npt.assert_raises(ValueError, fci.calc_inbag, X_train.shape[0],
+                      non_bootstrap_forest)
+
+
 
 def test_core_computation():
     inbag_ex = np.array([[1., 2., 0., 1.],
