### 2016-12-15 3682b8d
Implement correction for variance inflation when resampling without replacement
