+++ _downloads/plot_mpg.py
@@ -40,8 +40,8 @@ mpg_y_hat = mpg_forest.predict(mpg_X_test)
 
 # calculate inbag and unbiased variance
 mpg_inbag = fci.calc_inbag(mpg_X_train.shape[0], mpg_forest)
-mpg_V_IJ_unbiased = fci.random_forest_error(mpg_forest, mpg_inbag,
-                                            mpg_X_train, mpg_X_test)
+mpg_V_IJ_unbiased = fci.random_forest_error(mpg_forest, mpg_X_train,
+                                            mpg_X_test)
 
 # Plot error bars for predicted MPG using unbiased variance
 plt.errorbar(mpg_y_test, mpg_y_hat, yerr=np.sqrt(mpg_V_IJ_unbiased), fmt='o')
