### 2016-12-28 70f5622
improving _core_computation
