+++ forestci/forestci.py
@@ -57,14 +57,23 @@ def calc_inbag(n_samples, forest):
     return inbag
 
 
-def _core_computation(X_train, X_test, inbag, pred_centered, n_trees):
-    cov_hat = np.zeros((X_train.shape[0], X_test.shape[0]))
-
-    for t_idx in range(n_trees):
-        inbag_r = (inbag[:, t_idx] - 1).reshape(-1, 1)
-        pred_c_r = pred_centered.T[t_idx].reshape(1, -1)
-        cov_hat += np.dot(inbag_r, pred_c_r) / n_trees
-    V_IJ = np.sum(cov_hat ** 2, 0)
+def _core_computation(X_train, X_test, inbag, pred_centered, n_trees, memory_constrained=False, memory_limit=None):
+    if not memory_constrained:
+        return np.sum((np.dot(inbag-1,pred_centered.T)/n_trees)**2,0)
+    
+    if not memory_limit:
+        raise ValueError('If memory_constrained=True, must provide memory_limit.')
+    
+    chunk_size = int((memory_limit * 1e6) /(8.0 * X_train.shape[0]))
+    
+    if chunk_size==0:
+        min_limit = 8.0*X_train.shape[0]/(1e6)
+        raise ValueError('memory_limit provided is too small. For these dimensions, memory_limit must be greater than or equal to %.3e' % min_limit)
+    
+    chunk_edges = np.arange(0,X_test.shape[0]+chunk_size,chunk_size)
+    inds = range(X_test.shape[0])
+    chunks = [inds[chunk_edges[i]:chunk_edges[i+1]] for i in range(len(chunk_edges)-1)]
+    V_IJ = np.concatenate([np.sum((np.dot(inbag-1,pred_centered[chunk].T)/n_trees)**2,0) for chunk in chunks])
     return V_IJ
 
 
@@ -78,7 +87,7 @@ def _bias_correction(V_IJ, inbag, pred_centered, n_trees):
     return V_IJ_unbiased
 
 
-def random_forest_error(forest, X_train, X_test, inbag=None):
+def random_forest_error(forest, X_train, X_test, inbag=None, memory_constrained=False, memory_limit=None):
     """
     Calculates error bars from scikit-learn RandomForest estimators.
 
@@ -90,8 +99,11 @@ def random_forest_error(forest, X_train, X_test, inbag=None):
     forest : RandomForest
         Regressor or Classifier object.
 
-    X : ndarray
-        An array with shape (n_sample, n_features).
+    X_train : ndarray
+        An array with shape (n_train_sample, n_features).
+
+    X_test : ndarray
+        An array with shape (n_test_sample, n_features).
 
     inbag : ndarray (optional)
         The inbag matrix that fit the data. If set to `None` (default) it
@@ -99,6 +111,16 @@ def random_forest_error(forest, X_train, X_test, inbag=None):
         for which bootstrapping was set to `True`. That is, if sampling was
         done with replacement. Otherwise, users need to provide their own
         inbag matrix.
+
+    memory_constrained: boolean (optional)
+        Whether or not there is a restriction on memory. If False, it is 
+        assumed that a ndarry of shape (n_train_sample,n_test_sample) fits
+        in main memory. Setting to True can actually provide a speed up if
+        memory_limit is tuned to the optimal range.
+
+    memory_limit: int (optional)
+        An upper bound for how much memory the itermediate matrices will take
+        up in Megabytes. This must be provided if memory_constrained=True.
 
     Returns
     -------
@@ -125,7 +147,7 @@ def random_forest_error(forest, X_train, X_test, inbag=None):
     pred_mean = np.mean(pred, 0)
     pred_centered = pred - pred_mean
[…]
