### 2016-12-28 9f8699e
revised forestci to include calibrate functions --from Bryna
