### 2017-01-09 #124 @kpolimis
Hi Arfon,

Thanks for the note, happy new year to you as well. I apologize for not
being more responsive during the review process. We have started adding
additional computations from the original Wager et al. 2014 paper into the
package and I was focused on implementing those changes. I appreciate the
reviewer's time and will try to respond to their feedback within the next
two weeks.

Best,

Kivan

On Sat, Jan 7, 2017 at 7:32 PM, Arfon Smith <notifications@github.com>
wrote:

> 👋 @kpolimis <https://github.com/kpolimis>, happy new year. When do you
> think you might have a chance to respond/address to the reviewer's feedback
> here?
>
