### 2017-01-22 95e0c13
added forestci_revised with supplemental calculations from Wager et al. 2014
