### 2017-02-05 493442f
modified LICENSE, README, and paper
