### 2017-02-05 ac9abe8
modified LICENSE, README, and paper
