### 2017-02-22 bd41c42
revised calibrate.py
