### 2017-02-22 f4968ba
Add a notebook with some experiments around calibration code.
