### 2017-03-21 16558c1
Reinstate things that are on master, but got scuttled here for some reason.
