### 2017-03-27 e54f00e
modified test_forestci
