### 2017-03-27 146e1df
modified paper text and plots
