### 2017-03-27 #124 @kpolimis
Hi Arfon,

I've made the changes suggested by the reviewer in the paper and README.
The attached document contains the reviewer's original comments and my
responses are highlighted in yellow. Let me know if you have any questions.
Thanks for your support during this process.

Best,

Kivan

On Sun, Mar 12, 2017 at 12:52 PM, Arfon Smith <notifications@github.com>
wrote:

> 👋 @kpolimis <https://github.com/kpolimis> - have you had a chance to
> review this feedback yet?
>
