### 2017-03-27 #124 @pjotrp
The attachment is not visible. Also can you take a look at the LICENSE file. Is it really a BSD-3? Github should recognize the license contents to display it like in the icon on the top right of https://github.com/lomereiter/sambamba
