### 2017-03-27 62c0fd6
modified README and paper
