### 2017-03-27 e7c2723
modified README and paper
