+++ README.md
@@ -43,4 +43,7 @@ Tests are located in the `forestci\tests` folder and can be run with the `nosete
 
 ## Community guidelines
 
-To contribute to the software, report issues or problems with the software, or seek support
+Contributions to address any of the issues in the issue log are very welcome. Please submit improvements as pull requests against the repo after verifying that the existing tests pass and any new code is well covered by unit tests.
+
+Bug reports are also appreciated, please add them to the issue log after verifying that the issue does not already exist. Comments on existing issues are also welcome.
+
