### 2017-03-27 #124 @kpolimis
I've copied the BSD-3 clause License and changed the first line only, not
sure why the license is not being displayed like the repo you sent. Which
attachment are you speaking about?

On Mon, Mar 27, 2017 at 2:44 PM, Pjotr Prins <notifications@github.com>
wrote:

> The attachment is not visible. Also can you take a look at the LICENSE
> file. Is it really a BSD-3? Github should recognize the license contents to
> display it like in the icon on the top right of
> https://github.com/lomereiter/sambamba
>
