### 2017-03-27 2344cb8
modified example plots in paper
