### 2017-03-28 #124 @pjotrp
Thanks. It is probably not showing because the update is not on the master branch. The attachment you mentioned above with remarks in yellow...

The original software on which yours is based had an MIT license. I think it would be more appropriate to have that then to change it to BSD-3. Even so, the difference is minimal (BSD-3 forces you to show a copyright notice with a binary), so I think it is OK.
