### 2017-03-28 #124 @pjotrp
@DannyArends can you take a look at the improvements and state that you are happy with them? Use the following branch: https://github.com/kpolimis/forest-confidence-interval/tree/paper
