### 2017-03-28 a0a4195
modified paper text
