+++ README.md
@@ -12,7 +12,7 @@ random forest objects
 Compatible with Python2.7 and Python3.5
 
 This module is based on R code from Stefan Wager (see important links below)
-and is licensed under the BSD-3-Clause open source license (see [LICENSE](LICENSE))
+and is licensed under the MIT open source license (see [LICENSE](LICENSE))
 
 ## Important Links
 scikit-learn - http://scikit-learn.org/
