### 2017-03-28 #124 @kpolimis
Thanks for the feedback. I've updated the paper (attached) and revised the
License and README to reflect an MIT license. Lastly, my response to the
review is attached as a .docx. Let me know if you have any more questions.

Best,

Kivan

On Mon, Mar 27, 2017 at 6:26 PM, Pjotr Prins <notifications@github.com>
wrote:

> Thanks. It is probably not showing because the update is not on the master
> branch. The attachment you mentioned above with remarks in yellow...
>
> The original software on which yours is based had an MIT license. I think
> it would be more appropriate to have that then to change it to BSD-3. Even
> so, the difference is minimal (BSD-3 forces you to show a copyright notice
> with a binary), so I think it is OK.
>
