### 2017-06-05 #124 @DannyArends
Hey all,

I can look into the submission again this week.
However, I cannot see the attached response to reviewers _"Lastly, my response to the review is attached as a .docx"_

I will go through the issues I submitted and see if they are fixed

Gr,
Danny
