### 2017-06-05 #124 @DannyArends
Re-Review of scikit-learn-contrib/forest-confidence-interval

Still some issues with the version in the paper branch, also why haven't changes been pushed to the master branch of the scikit-learn-contrib repository ?

I have checked the boxes of the things which are now fixed, the remaining open issues are listed below.

**Performance:**
The software makes no performance claims, a performance comparison against the original R-code would be interesting to add. See scikit-learn-contrib/forest-confidence-interval#29

**A statement of need:**
There is no clear statement of need is not provided in the readme. See scikit-learn-contrib/forest-confidence-interval#30

**Example usage:**
The examples use two standard datasets from another package, however it is not explained in which format the users own data needs to be presented before it can be used. Please provide links to the relevant documentations.

The two examples do not show before and after plots of the dataset, as such it is not clear why the user should use this routine.

There is no description of what settings can be used in the calculation of inbag and unbiased variance and how these influence the results

See scikit-learn-contrib/forest-confidence-interval#31

**Functionality documentation:**
This package provides 2 functions to the user, which have been documented. However there are also 2 undocumented internal functions (_core_computation and _bias_correction), which should be documented. See scikit-learn-contrib/forest-confidence-interval#32

**Community guidelines:**
Please provide guidelines in the README file for others to:

    Seek support

See scikit-learn-contrib/forest-confidence-interval#34

**References:**
Both references (wager_confidence_2014 and wager_randomforestci_2016) are missing a DOI. See scikit-learn-contrib/forest-confidence-interval#36
