### 2017-06-07 #124 @pjotrp
@kpolimis I think these are good suggestions and are easy to fix. You don't have to state anything about performance. We prefer to see comments in the issue tracker or your source tree so they remain visible over time. Docx may not be the best format to use for sharing. Not everyone likes using Microsoft formats.
