### 2017-06-24 scikit-learn-contrib/forest-confidence-interval#36 @kpolimis
1. Wager et al. (2014) Confidence Intervals for Random Forests: The Jackknife and the Infinitesimal Jackknife does not have a DOI and is available on arXiv.
 - how do I include arXiv citations? This paper should be cited as arXiv:1311.4555
- https://arxiv.org/abs/1311.4555

2. wager_randomforestci_2016 is a GitHub repository and also does not have a DOI
- https://github.com/swager/randomForestCI
