### 2017-06-24 a1fa938
Add contribution statement to README.
