### 2017-06-24 97f022f
adding test for memory constrained _core_computation
