+++ _sources/auto_examples/plot_spam.rst.txt
@@ -29,7 +29,7 @@ of :func:`sklearn.datasets.make_classification`).
 
     import numpy as np
     from matplotlib import pyplot as plt
-    import sklearn.cross_validation as xval
+    from sklearn.model_selection import train_test_split
     from sklearn.ensemble import RandomForestClassifier
     import forestci as fci
     from sklearn.datasets import make_classification
@@ -37,7 +37,7 @@ of :func:`sklearn.datasets.make_classification`).
     spam_X, spam_y = make_classification(5000)
 
     # split the datainto training and test set
-    spam_X_train, spam_X_test, spam_y_train, spam_y_test = xval.train_test_split(
+    spam_X_train, spam_X_test, spam_y_train, spam_y_test = train_test_split(
                                                            spam_X, spam_y,
                                                            test_size=0.2)
 
@@ -68,7 +68,7 @@ of :func:`sklearn.datasets.make_classification`).
     plt.legend()
     plt.show()
 
-**Total running time of the script:** ( 2 minutes  38.414 seconds)
+**Total running time of the script:** ( 0 minutes  8.838 seconds)
 
 
 
@@ -87,4 +87,4 @@ of :func:`sklearn.datasets.make_classification`).
 
 .. rst-class:: sphx-glr-signature
 
-    `Generated by Sphinx-Gallery <http://sphinx-gallery.readthedocs.io>`_
+    `Generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
