+++ _sources/auto_examples/plot_mpg.rst.txt
@@ -66,7 +66,7 @@ various features of different cars, and their MPG.
     plt.ylabel('Predicted MPG')
     plt.show()
 
-**Total running time of the script:** ( 0 minutes  8.403 seconds)
+**Total running time of the script:** ( 0 minutes  7.912 seconds)
 
 
 
@@ -85,4 +85,4 @@ various features of different cars, and their MPG.
 
 .. rst-class:: sphx-glr-signature
 
-    `Generated by Sphinx-Gallery <http://sphinx-gallery.readthedocs.io>`_
+    `Generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
