### 2017-06-25 34e51ca
Delete calib.py
