### 2017-07-07 3b98031
Calibration with empirical Bayes.
This is the work of hzhao16 from #48, but without some large data files
that got added into the history along the way. Also, several added PEP8
fixes, and more comprehensive testing.
