+++ forestci/tests/test_forestci.py
@@ -25,8 +25,12 @@ def test_random_forest_error():
     forest = RandomForestRegressor(n_estimators=n_trees)
     forest.fit(X_train, y_train)
     inbag = fci.calc_inbag(X_train.shape[0], forest)
-    V_IJ_unbiased = fci.random_forest_error(forest, X_train, X_test)
-    npt.assert_equal(V_IJ_unbiased.shape[0], y_test.shape[0])
+    for ib in [inbag, None]:
+        for calibrate in [True, False]:
+            V_IJ_unbiased = fci.random_forest_error(forest, X_train, X_test,
+                                                    inbag=ib,
+                                                    calibrate=calibrate)
+        npt.assert_equal(V_IJ_unbiased.shape[0], y_test.shape[0])
 
     # We cannot calculate inbag from a non-bootstrapped forest. This is because
     # Scikit-learn trees do not store their own sample weights. If you did This
@@ -36,7 +40,6 @@ def test_random_forest_error():
 
     npt.assert_raises(ValueError, fci.calc_inbag, X_train.shape[0],
                       non_bootstrap_forest)
-
 
 
 def test_core_computation():
@@ -61,12 +64,14 @@ def test_core_computation():
 
     npt.assert_almost_equal(our_vij, r_vij)
 
-    our_vij = fci._core_computation(X_train_ex, X_test_ex, inbag_ex,
-                                    pred_centered_ex, n_trees,
-                                    memory_constrained=True, memory_limit=.01, 
-                                    test_mode=True)
+    for mc, ml in zip([True, False], [.01, None]):
+        our_vij = fci._core_computation(X_train_ex, X_test_ex, inbag_ex,
+                                        pred_centered_ex, n_trees,
+                                        memory_constrained=True,
+                                        memory_limit=.01,
+                                        test_mode=True)
 
-    npt.assert_almost_equal(our_vij, r_vij)    
+    npt.assert_almost_equal(our_vij, r_vij)
 
 
 def test_bias_correction():
@@ -91,3 +96,24 @@ def test_bias_correction():
                                             n_trees)
     r_unbiased_vij = np.array([-42.1875, 232.8125])
     npt.assert_almost_equal(our_vij_unbiased, r_unbiased_vij)
+
+
+def test_with_calibration():
+    # Test both with and without interpolation:
+    for n in [25, 205]:
+        X = np.random.rand(n).reshape(n // 5, 5)
+        y = np.random.rand(n // 5)
+
+        train_idx = np.arange(int(n // 5 * 0.75))
+        test_idx = np.arange(int(n//5 * 0.75), n//5)
+
+        y_test = y[test_idx]
+        y_train = y[train_idx]
+        X_test = X[test_idx]
+        X_train = X[train_idx]
+
+        n_trees = 4
+        forest = RandomForestRegressor(n_estimators=n_trees)
+        forest.fit(X_train, y_train)
+        V_IJ_unbiased = fci.random_forest_error(forest, X_train, X_test)
+        npt.assert_equal(V_IJ_unbiased.shape[0], y_test.shape[0])
