### 2017-07-07 02d850c
Get rid of non-ascii character (hopefully).
