+++ forestci/__init__.py
@@ -1,5 +1,8 @@
 from .forestci import (calc_inbag, random_forest_error,
-                          _core_computation, _bias_correction)
+                       _core_computation, _bias_correction)  # noqa
+
+from .calibration import *  # noqa
+
 from .version import __version__  # noqa
 
-__all__ = ["calc_inbag", "random_forest_error"]
+__all__ = ("calc_inbag", "random_forest_error")
