### 2017-07-07 2546621
Install pytest coverage.
