### 2017-07-07 e6b12ee
modified paper.md
