### 2017-07-07 scikit-learn-contrib/forest-confidence-interval#30 @kpolimis
added: "Our software is designed for individuals using `scikit-learn` random forest objects that want to add estimates of uncertainty to random forest predictors." to paper.md and README.md
