### 2017-07-07 scikit-learn-contrib/forest-confidence-interval#31 @kpolimis
1. Added: "The example data was imported as a matrix and converted to a data frame and separated into X (predictors) and y matrices (predictions). This module will work for matrices or `pandas` data frames.

2. Added before and after plots

3. Added: "The inbag matrix that fit the data is set to `None` by default, and the inbag will be inferred from the forest. However, this only works for trees for which bootstrapping was set to `True`. That is, if sampling was done with replacement. Otherwise, users need to provide their own inbag matrix."
