### 2017-07-07 79975d7
modified paper text
