### 2017-07-07 7cfa2dd
Back off of move to pytest. Let's stick to nose for now.
