+++ ci_scripts/test.sh
@@ -7,7 +7,7 @@ mkdir -p $TEST_DIR
 cd $TEST_DIR
 
 if [[ "$COVERAGE" == "true" ]]; then
-    py.test --pyargs $MODULE --cov-report term-missing --cov=$MODULE --doctest-modules
+    nosetests -s --with-coverage --cover-package=$MODULE $MODULE
 else
-  py.test --pyargs $MODULE
+    nosetests -s $MODULE
 fi
