### 2017-07-08 1b731b1
Use a much smaller default for the number of bins in EB fit.
