### 2017-07-08 12c4efa
Try relative import to get the calibration function.
