+++ forestci/__init__.py
@@ -1,8 +1,6 @@
 from .forestci import (calc_inbag, random_forest_error,
                        _core_computation, _bias_correction)  # noqa
 
-from .calibration import calibrateEB  # noqa
-
 from .version import __version__  # noqa
 
 __all__ = ("calc_inbag", "random_forest_error")
