### 2017-07-08 0c4a7c8
Increase the size, so that it tests with the calibration (doesn't just skip it).
