### 2017-07-08 ccc4301
Get rid of another non-ascii character.
