### 2017-07-08 bffd10b
These references have some non-ascii character.
