### 2017-07-08 8a8bc17
Use numpy infinity and sums.
