+++ _sources/auto_examples/plot_spam.rst.txt
@@ -68,7 +68,7 @@ of :func:`sklearn.datasets.make_classification`).
     plt.legend()
     plt.show()
 
-**Total running time of the script:** ( 0 minutes  9.034 seconds)
+**Total running time of the script:** ( 0 minutes  9.721 seconds)
 
 
 
