### 2017-07-27 50559de
Upgrade to use Python 3.6 on Travis.
Also add instructions for installation of dev version with 'pip', addresses #43
