+++ .travis.yml
@@ -15,7 +15,7 @@ env:
   matrix:
     - DISTRIB="conda" PYTHON_VERSION="2.7"
       NUMPY_VERSION="1.6.2" SCIPY_VERSION="0.11.0" CYTHON_VERSION="0.21"
-    - DISTRIB="conda" PYTHON_VERSION="3.5" COVERAGE="true"
+    - DISTRIB="conda" PYTHON_VERSION="3.6" COVERAGE="true"
       NUMPY_VERSION="1.10.4" SCIPY_VERSION="0.17.0" CYTHON_VERSION="0.23.4"
 
 install: source ci_scripts/install.sh
