### 2017-07-27 d76b9f2
Up the numpy requirement.
