+++ requirements.txt
@@ -1,3 +1,3 @@
-numpy>=1.6.1
+numpy>= 1.8.2
 nose>=1.1.2
 scikit-learn>=0.17
