### 2017-07-27 d6d60d9
Set minimal requirements for Travis in the travis config.
