### 2017-07-27 fd0e5a9
Bump scipy version.
