+++ .travis.yml
@@ -14,7 +14,7 @@ env:
     - MODULE=forestci
   matrix:
     - DISTRIB="conda" PYTHON_VERSION="2.7"
-      NUMPY_VERSION="1.8.1" SCIPY_VERSION="0.11.0" CYTHON_VERSION="0.21"
+      NUMPY_VERSION="1.8.1" SCIPY_VERSION="0.12.0" CYTHON_VERSION="0.21"
     - DISTRIB="conda" PYTHON_VERSION="3.6" COVERAGE="true"
       NUMPY_VERSION="1.10.4" SCIPY_VERSION="0.17.0" CYTHON_VERSION="0.23.4"
 
