### 2017-07-28 649ee11
Increment scipy version.
