### 2017-07-28 b8b486f
Bump scipy version.
