### 2017-07-28 ecb0aea
Bump scipy version
