### 2017-07-28 5fd5f0e
Bump numpy
