### 2017-07-28 d10e207
Bump versions. Don't bother testing old numpy/scipy
