### 2017-07-28 03122c5
Let conda decide what versions of numpy/scipy it wants.
