### 2017-07-28 scikit-learn-contrib/forest-confidence-interval#36 @arokem
Looks like neither of these references have a DOI. I am closing this, but please reopen if you think we are mistaken.
