### 2017-07-28 scikit-learn-contrib/forest-confidence-interval#34 @arokem
Closed through #46
