### 2017-07-28 scikit-learn-contrib/forest-confidence-interval#32 @arokem
These are private functions, that should not be used as part of the public API. I am not sure it would be a good idea to include them in the sphinx documentation.
