### 2017-07-28 scikit-learn-contrib/forest-confidence-interval#30 @arokem
Resolved through #47
