### 2017-07-28 scikit-learn-contrib/forest-confidence-interval#31 @arokem
@kpolimis : where are the commits that implement these changes? Please point to a Github link with these changes, so that I can merge them into master.
