+++ _sources/auto_examples/plot_mpg.rst.txt
@@ -66,7 +66,7 @@ various features of different cars, and their MPG.
     plt.ylabel('Predicted MPG')
     plt.show()
 
-**Total running time of the script:** ( 0 minutes  8.691 seconds)
+**Total running time of the script:** ( 0 minutes  8.554 seconds)
 
 
 
