### 2017-10-30 8036bb8
updated README and paper to include statement of need, revised community guidelines, API documentation, UCI machine learning citation, and emails for community support
