+++ paper/paper.bib
@@ -1,22 +1,32 @@
 
 @article{wager_confidence_2014,
-	title = {Confidence Intervals for Random Forests: The Jackknife and the Infinitesimal Jackknife},
+	title = {Confidence {Intervals} for {Random} {Forests}: {The} {Jackknife} and the {Infinitesimal} {Jackknife}},
 	volume = {15},
-	pages = {1625--1651},
+	issn = {1532-4435},
+	url = {http://dl.acm.org/citation.cfm?id=2627435.2638587},
 	number = {1},
-	journaltitle = {The Journal of Machine Learning Research},
+	journal = {J. Mach. Learn. Res.},
 	author = {Wager, Stefan and Hastie, Trevor and Efron, Bradley},
-	date = {2014-01},
-	file = {Wager2014.pdf:/Users/kivan/Library/Application Support/Zotero/Profiles/l980itek.default/zotero/storage/GGZPSFFS/Wager2014.pdf:application/pdf}
+	month = jan,
+	year = {2014},
+	keywords = {bagging, jackknife methods, Monte Carlo noise, variance estimation},
+	pages = {1625--1651}
 }
-	
-@online{wager_randomforestci_2016,
+
+@misc{wager_randomforestci_2016,
 	title = {{randomForestCI}},
 	url = {https://github.com/swager/randomForestCI},
-	abstract = {{randomForestCI}},
-	titleaddon = {{GitHub}},
-	author = {Wager, Stefan},
+	abstract = {randomForestCI},
 	urldate = {2016-09-23},
-	date = {2016-09},
-	file = {Snapshot:/Users/kivan/Library/Application Support/Zotero/Profiles/l980itek.default/zotero/storage/HKWBJCD2/randomForestCI.html:text/html}
+	author = {Wager, Stefan},
+	month = sep,
+	year = {2016}
+}
+
+@book{lichman_uci_2013,
+	title = {{UCI} {Machine} {Learning} {Repository}},
+	url = {http://archive.ics.uci.edu/ml},
+	publisher = {University of California, Irvine, School of Information and Computer Sciences},
+	author = {Lichman, M.},
+	year = {2013}
 }
