### 2017-10-30 #124 @kpolimis
tried to address outstanding review issues in the following PR: https://github.com/scikit-learn-contrib/forest-confidence-interval/pull/56
