### 2017-10-31 d80436a
rebased branch and edited README and paper for JOSS submission
