### 2017-10-31 9d28e99
fixed merge conflicts in README
