+++ README.md
@@ -8,7 +8,15 @@
 confidence intervals to scikit-learn random forest regression or classification
 objects. The core functions calculate an in-bag and error bars for
 random forest objects. Our software is designed for individuals using `scikit-learn` random forest objects that want to add estimates of uncertainty to random forest predictors.
+Forest algorithms are powerful
+[ensemble methods](http://scikit-learn.org/stable/modules/classes.html#module-sklearn.ensemble) for classification and regression. However, predictions from these
+algorithms do contain some amount of error.
 
+`forest-confidence-interval` is a Python module that adds a calculation of
+variance and computes confidence intervals to the basic functionality
+implemented in scikit-learn random forest regression or classification objects.
+The core functions here calculate an in-bag and error bars for random forest
+objects
 
 Compatible with Python2.7 and Python3.6
 
@@ -22,7 +30,7 @@ Stefan Wager's `randomForestCI` - https://github.com/swager/randomForestCI
 
 ## Installation and Usage
 Before installing the module you will need `numpy`, `scipy` and `scikit-learn`.  
-Dependencies associated with the previous modules may need root privileges to install
+Dependencies associated with the previous modules may need root privileges to install  
 Consult the [API Reference](http://contrib.scikit-learn.org/forest-confidence-interval/reference/index.html) for documentation on core functionality
 ```
 pip install numpy scipy scikit-learn
@@ -41,6 +49,13 @@ pip install forestci
 or, if you are installing from the source code:
 ```shell
 python setup.py install
+=======
+```
+
+If would like to install the development version of the software use:
+
+```shell
+pip install git+git://github.com/scikit-learn-contrib/forest-confidence-interval.git
 ```
 
 ## Why use `forest-confidence-interval`?
@@ -53,10 +68,16 @@ The regression example uses a popular UCI Machine Learning data set on cars whil
 [Examples gallery](http://contrib.scikit-learn.org/forest-confidence-interval/auto_examples/index.html)
 
 
-## Testing
-Requires installation of `nose` package. Tests are located in the `forestci/tests` folder and can be run with the `nosetests` command in the main directory
+## Contributing
 
-## Community guidelines
+Contributions are very welcome, but we ask that contributors abide by the
+[contributor covenant)[http://contributor-covenant.org/version/1/4/].
+
+To report issues with the software, please post to the
+[issue log](https://github.com/scikit-learn-contrib/forest-confidence-interval/issues)
+Bug reports are also appreciated, please add them to the issue log after
+verifying that the issue does not already exist.
+Comments on existing issues are also welcome.
 
 Contributions are very welcome, but we ask that contributors abide by the [contributor covenant](http://contributor-covenant.org/version/1/4/).
 
@@ -72,3 +93,7 @@ Please write code that complies with the Python style guide,
 [PEP8](https://www.python.org/dev/peps/pep-0008/)
 
 Please e-mail [Ariel Rokem](mailto:arokem@gmail.com), [Kivan Polimis](mailto:kivan.polimis@gmail.com), or [Bryna Hazelton](mailto:brynah@phys.washington.edu ) if you have any questions, suggestions or feedback.
+
+## Testing
+Requires installation of `nose` package. Tests are located in the `forestci/tests` folder
+and can be run with the `nosetests` command in the main directory
