### 2017-10-31 abd80ac
fixed more merge conflicts in README
