+++ README.md
@@ -4,17 +4,13 @@
 [![Coveralls Status](https://coveralls.io/repos/scikit-learn-contrib/forest-confidence-interval/badge.svg?branch=master&service=github)](https://coveralls.io/r/scikit-learn-contrib/forest-confidence-interval)
 [![CircleCI Status](https://circleci.com/gh/scikit-learn-contrib/forest-confidence-interval.svg?style=shield&circle-token=:circle-token)](https://circleci.com/gh/scikit-learn-contrib/forest-confidence-interval/tree/master)
 
-`forest-confidence-interval` is a Python module for calculating variance and adding
-confidence intervals to scikit-learn random forest regression or classification
-objects. The core functions calculate an in-bag and error bars for
-random forest objects. Our software is designed for individuals using `scikit-learn` random forest objects that want to add estimates of uncertainty to random forest predictors.
 Forest algorithms are powerful
 [ensemble methods](http://scikit-learn.org/stable/modules/classes.html#module-sklearn.ensemble) for classification and regression. However, predictions from these
 algorithms do contain some amount of error.
 
 `forest-confidence-interval` is a Python module that adds a calculation of
 variance and computes confidence intervals to the basic functionality
-implemented in scikit-learn random forest regression or classification objects.
+implemented in `scikit-learn` random forest regression or classification objects.
 The core functions here calculate an in-bag and error bars for random forest
 objects
 
@@ -30,8 +26,9 @@ Stefan Wager's `randomForestCI` - https://github.com/swager/randomForestCI
 
 ## Installation and Usage
 Before installing the module you will need `numpy`, `scipy` and `scikit-learn`.  
-Dependencies associated with the previous modules may need root privileges to install  
+Dependencies associated with the previous modules may need root privileges to install
 Consult the [API Reference](http://contrib.scikit-learn.org/forest-confidence-interval/reference/index.html) for documentation on core functionality
+
 ```
 pip install numpy scipy scikit-learn
 ```
@@ -49,7 +46,6 @@ pip install forestci
 or, if you are installing from the source code:
 ```shell
 python setup.py install
-=======
 ```
 
 If would like to install the development version of the software use:
@@ -58,28 +54,13 @@ If would like to install the development version of the software use:
 pip install git+git://github.com/scikit-learn-contrib/forest-confidence-interval.git
 ```
 
-## Why use `forest-confidence-interval`?
-The variability in predictions is important for measuring standard errors and estimating standard errors. `forest-confidence-interval` is a Python module for calculating variance and adding confidence intervals to the popular python library `scikit-learn`. The software is compatible with both `scikit-learn` random forest regression or classification objects. Our software is designed for individuals using `scikit-learn` random forest objects that want to add estimates of uncertainty to random forest predictors.
-
-
 ## Examples
-The examples (gallery below) demonstrates the package functionality with random forest classifiers and regression models.
-The regression example uses a popular UCI Machine Learning data set on cars while the classifier example demonstrates how to add measurements of uncertainty to predictions of spam emails.
-[Examples gallery](http://contrib.scikit-learn.org/forest-confidence-interval/auto_examples/index.html)
-
+See [examples gallery](http://contrib.scikit-learn.org/forest-confidence-interval/auto_examples/index.html)
 
 ## Contributing
 
 Contributions are very welcome, but we ask that contributors abide by the
-[contributor covenant)[http://contributor-covenant.org/version/1/4/].
-
-To report issues with the software, please post to the
-[issue log](https://github.com/scikit-learn-contrib/forest-confidence-interval/issues)
-Bug reports are also appreciated, please add them to the issue log after
[…]
