### 2017-10-31 5a1e71b
README merge conflicts
