### 2017-10-31 1c23957
modified README, paper text and bib for JOSS submission
