+++ README.md
@@ -12,7 +12,7 @@ algorithms do contain some amount of error.
 variance and computes confidence intervals to the basic functionality
 implemented in scikit-learn random forest regression or classification objects.
 The core functions here calculate an in-bag and error bars for random forest
-objects
+objects.
 
 Compatible with Python2.7 and Python3.6
 
@@ -25,8 +25,10 @@ scikit-learn - http://scikit-learn.org/
 Stefan Wager's `randomForestCI` - https://github.com/swager/randomForestCI
 
 ## Installation and Usage
-Before installing the module you will need `numpy`, `scipy` and `scikit-learn`.  
+Before installing the module you will need `numpy`, `scipy` and `scikit-learn`.
 Dependencies associated with the previous modules may need root privileges to install
+Consult the [API Reference](http://contrib.scikit-learn.org/forest-confidence-interval/reference/index.html) for documentation on core functionality
+
 ```
 pip install numpy scipy scikit-learn
 ```
@@ -51,14 +53,20 @@ If would like to install the development version of the software use:
 ```shell
 pip install git+git://github.com/scikit-learn-contrib/forest-confidence-interval.git
 ```
+## Why use `forest-confidence-interval`?
+The variability in predictions is important for measuring standard errors and estimating standard errors. `forest-confidence-interval` is a Python module for calculating variance and adding confidence intervals to the popular python library `scikit-learn`. The software is compatible with both `scikit-learn` random forest regression or classification objects. Our software is designed for individuals using `scikit-learn` random forest objects that want to add estimates of uncertainty to random forest predictors.
 
 ## Examples
-See [examples gallery](http://contrib.scikit-learn.org/forest-confidence-interval/auto_examples/index.html)
+
+The examples (gallery below) demonstrates the package functionality with random forest classifiers and regression models.
+The regression example uses a popular UCI Machine Learning data set on cars while the classifier example demonstrates how to add measurements of uncertainty to predictions of spam emails.
+
+[Examples gallery](http://contrib.scikit-learn.org/forest-confidence-interval/auto_examples/index.html)
 
 ## Contributing
 
 Contributions are very welcome, but we ask that contributors abide by the
-[contributor covenant)[http://contributor-covenant.org/version/1/4/].
+[contributor covenant]([http://contributor-covenant.org/version/1/4/).
 
 To report issues with the software, please post to the
 [issue log](https://github.com/scikit-learn-contrib/forest-confidence-interval/issues)
@@ -69,8 +77,10 @@ Comments on existing issues are also welcome.
 Please submit improvements as pull requests against the repo after verifying
 that the existing tests pass and any new code is well covered by unit tests.
 Please write code that complies with the Python style guide,
-[PEP8](https://www.python.org/dev/peps/pep-0008/)
+[PEP8](https://www.python.org/dev/peps/pep-0008/).
+
+E-mail [Ariel Rokem](mailto:arokem@gmail.com), [Kivan Polimis](mailto:kivan.polimis@gmail.com), or [Bryna Hazelton](mailto:brynah@phys.washington.edu ) if you have any questions, suggestions or feedback.
 
 ## Testing
 Requires installation of `nose` package. Tests are located in the `forestci/tests` folder
-and can be run with the `nosetests` command in the main directory
+and can be run with the `nosetests` command in the main directory.
