### 2017-10-31 #124 @kpolimis
needed to rebase and created new PR for review: https://github.com/scikit-learn-contrib/forest-confidence-interval/pull/58
