+++ _sources/auto_examples/plot_spam.rst.txt
@@ -68,11 +68,13 @@ of :func:`sklearn.datasets.make_classification`).
     plt.legend()
     plt.show()
 
-**Total running time of the script:** ( 0 minutes  9.340 seconds)
+**Total running time of the script:** ( 0 minutes  9.953 seconds)
 
 
 
-.. container:: sphx-glr-footer
+.. only :: html
+
+ .. container:: sphx-glr-footer
 
 
   .. container:: sphx-glr-download
@@ -85,6 +87,9 @@ of :func:`sklearn.datasets.make_classification`).
 
      :download:`Download Jupyter notebook: plot_spam.ipynb <plot_spam.ipynb>`
 
-.. rst-class:: sphx-glr-signature
 
-    `Generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+.. only:: html
+
+ .. rst-class:: sphx-glr-signature
+
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
