+++ _sources/auto_examples/index.rst.txt
@@ -60,7 +60,9 @@ of two categories (e.g., spam/non-spam) based on a number of features.
 
 
 
-.. container:: sphx-glr-footer
+.. only :: html
+
+ .. container:: sphx-glr-footer
 
 
   .. container:: sphx-glr-download
@@ -73,6 +75,9 @@ of two categories (e.g., spam/non-spam) based on a number of features.
 
     :download:`Download all examples in Jupyter notebooks: auto_examples_jupyter.zip <//home/ubuntu/forest-confidence-interval/doc/auto_examples/auto_examples_jupyter.zip>`
 
-.. rst-class:: sphx-glr-signature
 
-    `Generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+.. only:: html
+
+ .. rst-class:: sphx-glr-signature
+
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
