+++ _sources/auto_examples/plot_mpg.rst.txt
@@ -66,11 +66,13 @@ various features of different cars, and their MPG.
     plt.ylabel('Predicted MPG')
     plt.show()
 
-**Total running time of the script:** ( 0 minutes  8.554 seconds)
+**Total running time of the script:** ( 0 minutes  7.293 seconds)
 
 
 
-.. container:: sphx-glr-footer
+.. only :: html
+
+ .. container:: sphx-glr-footer
 
 
   .. container:: sphx-glr-download
@@ -83,6 +85,9 @@ various features of different cars, and their MPG.
 
      :download:`Download Jupyter notebook: plot_mpg.ipynb <plot_mpg.ipynb>`
 
-.. rst-class:: sphx-glr-signature
 
-    `Generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
+.. only:: html
+
+ .. rst-class:: sphx-glr-signature
+
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
