### 2017-11-04 #124 @pjotrp
@DannyArends can you take a final look?
