### 2017-11-07 #124 @DannyArends
Re-Re-Review of scikit-learn-contrib/forest-confidence-interval

I have checked the boxes of the things which are now fixed, but again there are some (6) remaining open major issues which according to me should be fixed before publication. Additionally 1 minor issue which may be ignored by the authors.

Major issues:

**Example usage:**
1) One of the examples use a data set from the UC Irvine Machine Learning Repository, however the data set used was not produced by UC Irvine Machine Learning Repository, so please cite the paper which collected the data used:

"Quinlan,R. (1993). Combining Instance-Based and Model-Based Learning. In Proceedings on the Tenth International Conference of Machine Learning, 236-243."

In the paper this data set is WRONGLY attributed, please update the citation in the paper as well.

2) Again, the examples do not show **before** and **after** plots of the data set, as such it is not clear why the user should use this routine.

3) Additionally, there is no description of what settings can be used in the calculation of inbag and unbiased variance and how these influence the results. Additionally no explanation is given why the Y-error bars need to be SQRT transformed before plotting. The examples provided are meaningless to me (and users) without better documentation and explanation of the code used in the examples.

**Functionality documentation:**
4) As mentioned before only 1/2 of the functions have documentation, the internal functions are left undocumented for some reason beyond my comprehension. Please document all of the functions in the code as required by the JOSS journal

**References:**
5) A reference to the CAR MPG dataset is WRONGLY attributed to: @llichman_uci_2013

The UCI page that you link to provides the correct citation for the dataset:

"Quinlan,R. (1993). Combining Instance-Based and Model-Based Learning. In Proceedings on the Tenth International Conference of Machine Learning, 236-243"

6) There are no DOIs in the bib file belonging to the article, this is a requirement of the JOSS journal

Minor issue:

**Performance:**
The software makes no performance claims, a performance comparison against the original R-code would be interesting to add. See scikit-learn-contrib/forest-confidence-interval#29
