### 2017-11-07 #124 @DannyArends
Dear @arokem

"Does this PR address this issue? scikit-learn-contrib/forest-confidence-interval#62"
- Yes, that is enough.

However why is there a sqrt in the example code ?

line 59: np.sqrt(spam_V_IJ_unbiased[idx_ham])

"Wasn't this addressed in this PR: scikit-learn-contrib/forest-confidence-interval#59?"

- Yes it was addressed, that's why I removed it from the review list, and removed it from my comment. However I now realize that the email send might have contained the old text. Since I copy pasted it and removed the issues that were fixed.
