### 2017-11-08 2ef84b5
modified paper to answer latest JOSS review request
