+++ paper/paper.bib
@@ -10,7 +10,8 @@
 	month = jan,
 	year = {2014},
 	keywords = {bagging, jackknife methods, Monte Carlo noise, variance estimation},
-	pages = {1625--1651}
+	pages = {1625--1651},
+	file = {Wageretal2014.pdf:C\:\\Users\\Kivan\\Zotero\\storage\\7ZBXYKH2\\Wageretal2014.pdf:application/pdf}
 }
 
 @misc{wager_randomforestci_2016,
@@ -23,10 +24,16 @@
 	year = {2016}
 }
 
-@book{lichman_uci_2013,
-	title = {{UCI} {Machine} {Learning} {Repository}},
-	url = {http://archive.ics.uci.edu/ml},
-	publisher = {University of California, Irvine, School of Information and Computer Sciences},
-	author = {Lichman, M.},
-	year = {2013}
+@inproceedings{quinlan_combining_1993,
+	address = {San Francisco, CA, USA},
+	series = {{ICML}'93},
+	title = {Combining {Instance}-based and {Model}-based {Learning}},
+	isbn = {1-55860-307-7},
+	url = {http://dl.acm.org/citation.cfm?id=3091529.3091560},
+	booktitle = {Proceedings of the {Tenth} {International} {Conference} on {International} {Conference} on {Machine} {Learning}},
+	publisher = {Morgan Kaufmann Publishers Inc.},
+	author = {Quinlan, J. Ross},
+	year = {1993},
+	pages = {236--243},
+	file = {Quinlan1993.pdf:C\:\\Users\\Kivan\\Zotero\\storage\\XNZAZBY7\\Quinlan1993.pdf:application/pdf}
 }
