### 2017-11-08 777659f
made syntax changes to paper
