### 2017-11-08 #124 @kpolimis
Re: "Additionally, there is no description of what settings can be used in the calculation of inbag and unbiased variance and how these influence the results. Additionally no explanation is given why the Y-error bars need to be SQRT transformed before plotting. The examples provided are meaningless to me (and users) without better documentation and explanation of the code used in the examples."

plotting with standard deviations addressed in: https://github.com/scikit-learn-contrib/forest-confidence-interval/pull/63
