### 2017-11-08 #124 @DannyArends
DOIs are still missing from the citations, but all in all I'm Happy.

I checked all the remaining boxes.
