### 2017-11-09 #124 @arfon
@kpolimis - could you please merge this PR https://github.com/scikit-learn-contrib/forest-confidence-interval/pull/64 and then make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
