+++ README.md
@@ -5,13 +5,13 @@
 [![CircleCI Status](https://circleci.com/gh/scikit-learn-contrib/forest-confidence-interval.svg?style=shield&circle-token=:circle-token)](https://circleci.com/gh/scikit-learn-contrib/forest-confidence-interval/tree/master)
 
 Forest algorithms are powerful
-[ensemble methods](http://scikit-learn.org/stable/modules/classes.html#module-sklearn.ensemble) for classification and regression. However, predictions from these
-algorithms do contain some amount of error.
+[ensemble methods](http://scikit-learn.org/stable/modules/classes.html#module-sklearn.ensemble) for classification and regression. However, predictions from these algorithms do contain some amount of error. Prediction variability can illustrate how influential
+the training set is for producing the observed random forest predictions.
 
 `forest-confidence-interval` is a Python module that adds a calculation of
 variance and computes confidence intervals to the basic functionality
 implemented in scikit-learn random forest regression or classification objects.
-The core functions here calculate an in-bag and error bars for random forest
+The core functions calculate an in-bag and error bars for random forest
 objects.
 
 Compatible with Python2.7 and Python3.6
@@ -20,9 +20,9 @@ This module is based on R code from Stefan Wager (see important links below)
 and is licensed under the MIT open source license (see [LICENSE](LICENSE))
 
 ## Important Links
-scikit-learn - http://scikit-learn.org/
+`scikit-learn` - http://scikit-learn.org/
 
-Stefan Wager's `randomForestCI` - https://github.com/swager/randomForestCI
+Stefan Wager's `randomForestCI` - https://github.com/swager/randomForestCI (deprecated in favor of `grf`: https://github.com/swager/grf)
 
 ## Installation and Usage
 Before installing the module you will need `numpy`, `scipy` and `scikit-learn`.  
@@ -54,12 +54,12 @@ If would like to install the development version of the software use:
 pip install git+git://github.com/scikit-learn-contrib/forest-confidence-interval.git
 ```
 ## Why use `forest-confidence-interval`?
-The variability in predictions is important for measuring and estimating standard errors. `forest-confidence-interval` is a Python module for calculating variance and adding confidence intervals to the popular python library `scikit-learn`. The software is compatible with both `scikit-learn` random forest regression or classification objects. Our software is designed for individuals using `scikit-learn` random forest objects that want to add estimates of uncertainty to random forest predictors.
+Our software is designed for individuals using `scikit-learn` random forest objects that want to add estimates of uncertainty to random forest predictors. Prediction variability demonstrates how much the training set influences results and is important for estimating standard errors. `forest-confidence-interval` is a Python module for calculating variance and adding confidence intervals to the popular Python library `scikit-learn`. The software is compatible with both `scikit-learn` random forest regression or classification objects.
 
 ## Examples
 
 The examples (gallery below) demonstrates the package functionality with random forest classifiers and regression models.
-The regression example uses a popular UCI Machine Learning data set on cars while the classifier example demonstrates how to add measurements of uncertainty to predictions of spam emails.
+The regression example uses a popular UCI Machine Learning data set on cars while the classifier example simulates how to add measurements of uncertainty to tasks like predicting spam emails.
 
 [Examples gallery](http://contrib.scikit-learn.org/forest-confidence-interval/auto_examples/index.html)
 
