+++ paper/paper.bib
@@ -5,13 +5,12 @@
 	issn = {1532-4435},
 	url = {http://dl.acm.org/citation.cfm?id=2627435.2638587},
 	number = {1},
-	journal = {J. Mach. Learn. Res.},
+	journal = {Journal of Machine Learning Research},
 	author = {Wager, Stefan and Hastie, Trevor and Efron, Bradley},
 	month = jan,
 	year = {2014},
 	keywords = {bagging, jackknife methods, Monte Carlo noise, variance estimation},
-	pages = {1625--1651},
-	file = {Wageretal2014.pdf:C\:\\Users\\Kivan\\Zotero\\storage\\7ZBXYKH2\\Wageretal2014.pdf:application/pdf}
+	pages = {1625--1651}
 }
 
 @misc{wager_randomforestci_2016,
@@ -34,6 +33,5 @@
 	publisher = {Morgan Kaufmann Publishers Inc.},
 	author = {Quinlan, J. Ross},
 	year = {1993},
-	pages = {236--243},
-	file = {Quinlan1993.pdf:C\:\\Users\\Kivan\\Zotero\\storage\\XNZAZBY7\\Quinlan1993.pdf:application/pdf}
+	pages = {236--243}
 }
