+++ _sources/auto_examples/plot_spam.rst.txt
@@ -17,8 +17,18 @@ of :func:`sklearn.datasets.make_classification`).
 
 
 
-.. image:: /auto_examples/images/sphx_glr_plot_spam_001.png
-    :align: center
+.. rst-class:: sphx-glr-horizontal
+
+
+    *
+
+      .. image:: /auto_examples/images/sphx_glr_plot_spam_001.png
+            :scale: 47
+
+    *
+
+      .. image:: /auto_examples/images/sphx_glr_plot_spam_002.png
+            :scale: 47
 
 
 
@@ -48,6 +58,17 @@ of :func:`sklearn.datasets.make_classification`).
     spam_RFC.fit(spam_X_train, spam_y_train)
     spam_y_hat = spam_RFC.predict_proba(spam_X_test)
 
+    idx_spam = np.where(spam_y_test == 1)[0]
+    idx_ham = np.where(spam_y_test == 0)[0]
+
+    # Histogram predictions without error bars:
+    fig, ax = plt.subplots(1)
+    ax.hist(spam_y_hat[idx_spam, 1], histtype='step', label='spam')
+    ax.hist(spam_y_hat[idx_ham, 1], histtype='step', label='not spam')
+    ax.set_xlabel('Prediction (spam probability)')
+    ax.set_ylabel('Number of observations')
+    plt.legend()
+
     # calculate inbag and unbiased variance
     spam_inbag = fci.calc_inbag(spam_X_train.shape[0], spam_RFC)
     spam_V_IJ_unbiased = fci.random_forest_error(spam_RFC, spam_X_train,
@@ -55,20 +76,21 @@ of :func:`sklearn.datasets.make_classification`).
 
     # Plot forest prediction for emails and standard deviation for estimates
     # Blue points are spam emails; Green points are non-spam emails
-    idx = np.where(spam_y_test == 1)[0]
-    plt.errorbar(spam_y_hat[idx, 1], np.sqrt(spam_V_IJ_unbiased[idx]),
-                 fmt='.', alpha=0.75, label='spam')
+    fig, ax = plt.subplots(1)
+    ax.scatter(spam_y_hat[idx_spam, 1],
+               np.sqrt(spam_V_IJ_unbiased[idx_spam]),
+               label='spam')
 
-    idx = np.where(spam_y_test == 0)[0]
-    plt.errorbar(spam_y_hat[idx, 1], np.sqrt(spam_V_IJ_unbiased[idx]),
-                 fmt='.', alpha=0.75, label='not spam')
+    ax.scatter(spam_y_hat[idx_ham, 1],
+               np.sqrt(spam_V_IJ_unbiased[idx_ham]),
+               label='not spam')
 
-    plt.xlabel('Prediction (spam probability)')
-    plt.ylabel('Standard deviation')
+    ax.set_xlabel('Prediction (spam probability)')
+    ax.set_ylabel('Standard deviation')
     plt.legend()
     plt.show()
 
-**Total running time of the script:** ( 0 minutes  6.946 seconds)
+**Total running time of the script:** ( 0 minutes  7.640 seconds)
 
 
 
