+++ _sources/auto_examples/plot_mpg.rst.txt
@@ -73,7 +73,7 @@ various features of different cars, and their MPG.
     plt.ylabel('Predicted MPG')
     plt.show()
 
-**Total running time of the script:** ( 0 minutes  7.050 seconds)
+**Total running time of the script:** ( 0 minutes  7.394 seconds)
 
 
 
