### 2017-11-09 8591ae1
modified examples code
