### 2017-11-09 #124 @kpolimis
@arfon I merged PR #64 and created a DOI of the archive with Zenodo (10.5281/zenodo.1044209)
