### 2017-11-09 #124 @kpolimis
Hi Arfon,

I just checked the version of the paper that appears online (
http://www.theoj.org/joss-papers/joss.00124/10.21105.joss.00124.pdf) and it
is a previous edition of the paper and not the version from Zenodo (
https://zenodo.org/record/1044209#.WgTE-GhSzD4). Is it possible to update
the JOSS paper to the latest version on master/Zenodo?

Thanks,

Kivan

Kivan Polimis
Postdoctoral Researcher
Università Commerciale Luigi Bocconi
Carlo F. Dondena Centre for Research on Social Dynamics and Public Policy
address: 5-a2-04, Via Guglielmö Röntgen 1 - 20136 Milano, Italy
<https://www.google.com/maps/dir//Via+Guglielm%C3%B6+R%C3%B6ntgen,+1,+20136+Milano+MI,+Italy/@45.4507127,9.1852656,17z/data=!4m17!1m7!3m6!1s0x4786c404480a59b9:0xe5f0cfe57a3877b!2sVia+Guglielm%C3%B6+R%C3%B6ntgen,+1,+20136+Milano+MI,+Italy!3b1!8m2!3d45.4507127!4d9.1874543!4m8!1m0!1m5!1m1!1s0x4786c404480a59b9:0xe5f0cfe57a3877b!2m2!1d9.1874543!2d45.4507127!3e3?hl=en>
office: +39.02.5836.5378
email: kivan.polimis@unibocconi.it
kivanpolimis.com

On Thu, Nov 9, 2017 at 8:21 AM, Arfon Smith <notifications@github.com>
wrote:

> Closed #124 <https://github.com/openjournals/joss-reviews/issues/124>.
>
